requirements.txt
streamlit>=1.25
pandas>=2.0
python-docx>=0.8.11
docx2pdf>=0.1.8
openpyxl>=3.1.2
